Got rid of TESTEntrant class
