+++ app/src/main/java/com/example/pygmyhippo/organizer/EntrantArrayAdapter.java
@@ -19,7 +19,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.example.pygmyhippo.common.Entrant;
-import com.example.pygmyhippo.common.TESTEntrant;
 
 import java.util.ArrayList;
 
@@ -28,11 +27,10 @@ import java.util.ArrayList;
  * @author Kori
  * TODO:
  *  - Make profile image change based on the user profile
- *  - Change TESTEntrant to Entrant
  */
-public class EntrantArrayAdapter extends ArrayAdapter<TESTEntrant> {
+public class EntrantArrayAdapter extends ArrayAdapter<Entrant> {
 
-    public EntrantArrayAdapter(Context context, ArrayList<TESTEntrant> entrantListData) {
+    public EntrantArrayAdapter(Context context, ArrayList<Entrant> entrantListData) {
         super(context, 0, entrantListData);
     }
 
@@ -57,8 +55,7 @@ public class EntrantArrayAdapter extends ArrayAdapter<TESTEntrant> {
             view = convertView;
         }
 
-        // FIXME: Change when test entrant is deleted
-        TESTEntrant entrant = getItem(position);
+        Entrant entrant = getItem(position);
 
         // Get all the views from the entrants_content in our xml file
         TextView accountName = view.findViewById(R.id.o_entrant_name_view);
